+++ Infrastructure.Identity/Contexts/IdentityContext.cs
@@ -1,0 +1,16 @@
+﻿using Infrastructure.Identity.Models;
+using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
+using Microsoft.EntityFrameworkCore;
+using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace Infrastructure.Identity.Contexts
+{
+    public class IdentityContext : IdentityDbContext<ApplicationUser>
+    {
+        public IdentityContext(DbContextOptions<IdentityContext> options) : base(options)
+        {
+        }
+    }
+}