+++ Infrastructure.Identity/ServiceRegistration.cs
@@ -1,10 +1,21 @@
-﻿using System;
+﻿using Infrastructure.Identity.Contexts;
+using Microsoft.EntityFrameworkCore;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+using System;
 using System.Collections.Generic;
 using System.Text;
 
 namespace Infrastructure.Identity
 {
-    class ServiceRegistration
+    public static class ServiceRegistration
     {
+        public static void AddIdentityInfrastructure(this IServiceCollection services, IConfiguration configuration)
+        {
+            services.AddDbContext<IdentityContext>(options =>
+                options.UseSqlServer(
+                    configuration.GetConnectionString("IdentityConnection"),
+                    b => b.MigrationsAssembly(typeof(IdentityContext).Assembly.FullName)));
+        }
     }
 }